+++ v2/.gitignore
@@ -1,3 +1,2 @@
-old/*
-__pycache__/*
-.pytest_cache/*
+backend/__pycache__/*
+backend/.pytest_cache/*
