+++ v2/backend/visit.py
@@ -3,10 +3,9 @@ from typing import Tuple
 
 from returns.result import Result, Failure, Success
 
-from apartment import ApartmentDatabase
-import errors
-from errors import Error
-from util import get_time, get_date
+from backend.apartment import ApartmentDatabase
+from backend.errors import *
+from backend.util import get_time, get_date
 
 class VisitorManager:
     def __init__(self, apts: ApartmentDatabase, path: str):
@@ -23,12 +22,13 @@ class VisitorManager:
     def is_visitor(self, number: int, name: str) -> bool:
         return (number, name) in self.visitors
 
+    # TODO: write to the file on sign-in
     def sign_in(self, number: int, name: str) -> Result[None, Error]:
         match self.apts.add_visitor(number, name):
             case Success(_):
                 self.add_visitor(number, name)
                 self.apts.save()
-            case Failure(errors.DuplicateVisitor()):
+            case Failure(DuplicateVisitor()):
                 self.add_visitor(number, name)
             case Failure(x):
                 return Failure(x)
@@ -40,7 +40,7 @@ class VisitorManager:
             self.visitors.pop((number, name))
             return Success(None)
         else:
-            return Failure(errors.VisitorNotFound(f"Could not find and sign-out visitor '{name}'"))
+            return Failure(VisitorNotFound(f"Could not find and sign-out visitor '{name}'"))
 
     # could happen that the name has a comma in it and this breaks, but frontend will filter that :)
     def add_log(self, text):
