+++ v2/data/apt.json
@@ -1,1 +1,1 @@
-[{"number": 69, "tenants": ["alice", "bob", "cindy"], "visitors": ["david", "ethan", "frank", "bruh"]}, {"number": 420, "tenants": ["alex", "ben", "catherine"], "visitors": ["daniel", "elise", "francis"]}, {"number": 101, "tenants": ["Tanya"], "visitors": ["John Watson", "Mary Watson"]}, {"number": 102, "tenants": ["Mike Bowen"], "visitors": ["Tom"]}, {"number": 103, "tenants": ["Gail Lee"], "visitors": []}, {"number": 104, "tenants": [], "visitors": []}, {"number": 105, "tenants": [], "visitors": []}, {"number": 201, "tenants": ["john", "mary"], "visitors": ["Jerry", "Eric Wang", "Tom"]}, {"number": 202, "tenants": [], "visitors": []}, {"number": 203, "tenants": [], "visitors": []}, {"number": 204, "tenants": [], "visitors": []}, {"number": 205, "tenants": [], "visitors": ["Barack Obama"]}, {"number": 301, "tenants": [], "visitors": []}, {"number": 302, "tenants": [], "visitors": []}, {"number": 303, "tenants": [], "visitors": []}, {"number": 304, "tenants": [], "visitors": []}, {"number": 305, "tenants": [], "visitors": []}]
+[{"number": 69, "tenants": ["alice", "bob", "cindy"], "visitors": ["david", "ethan", "frank", "bruh"]}, {"number": 420, "tenants": ["alex", "ben", "catherine"], "visitors": ["daniel", "elise", "francis"]}, {"number": 101, "tenants": ["Tanya"], "visitors": ["John Watson", "Mary Watson"]}, {"number": 102, "tenants": ["Mike Bowen"], "visitors": ["Tom", "david andrews"]}, {"number": 103, "tenants": ["Gail Lee"], "visitors": []}, {"number": 104, "tenants": [], "visitors": []}, {"number": 105, "tenants": [], "visitors": []}, {"number": 201, "tenants": ["john", "mary"], "visitors": ["Jerry", "Eric Wang", "Tom"]}, {"number": 202, "tenants": [], "visitors": []}, {"number": 203, "tenants": [], "visitors": []}, {"number": 204, "tenants": [], "visitors": []}, {"number": 205, "tenants": [], "visitors": ["Barack Obama"]}, {"number": 301, "tenants": [], "visitors": []}, {"number": 302, "tenants": [], "visitors": []}, {"number": 303, "tenants": [], "visitors": []}, {"number": 304, "tenants": [], "visitors": []}, {"number": 305, "tenants": [], "visitors": []}]
