+++ v2/visitor-logs/2022-07-06.csv
@@ -1,2 +1,5 @@
 69,bruh,13:30,13:30
 69,bruh,20:40,20:40
+102,david andrews,21:19,21:19
+69,bruh,21:45,21:45
+69,bruh,21:46,21:46
